refactor/메인페이지 api 파일 분리
